implement redraw requests
