+++ src/main.rs
@@ -8,6 +8,7 @@ mod logger;
 mod colors;
 mod shapes;
 mod util;
+mod camera;
 
 use shapes::get_shape;
 use util::*;
@@ -102,151 +103,98 @@ fn main() {
                 }
                 _ => return,
             }
+            glutin::event::Event::RedrawRequested(_) => {
+                let mut target = display.draw();
+                target.clear_color_and_depth((0.0, 0.0, 1.0, 1.0), 1.0);
+                let uniforms = uniform! {
+                    matrix: [
+                        [0.5, 0.0, 0.0, 0.0],
+                        [0.0, 0.5, 0.0, 0.0],
+                        [0.0, 0.0, 0.5, 0.0],
+                        [0.5, 0.0, 0.0, 1.0f32],
+                    ]
+                };
+        
+                let texture_uniforms = uniform! {
+                    matrix: [
+                        [0.5, 0.0, 0.0, 0.0],
+                        [0.0, 0.5, 0.0, 0.0],
+                        [0.0, 0.0, 0.5, 0.0],
+                        [-0.5, 0.0, 0.0, 1.0f32],
+                    ],
+                    tex: &example_image
+                };
+        
+                let model = [
+                    [0.0,    0.0,  -1.0,  0.0], // X-axis rotation (no change to X-axis values)
+                    [0.0,    1.0,   0.0,  0.0], // Y-axis rotation (no change to Y-axis values)
+                    [1.0,    0.0,   0.0,  0.0], // Z-axis rotation (no change to Z-axis values)
+                    [0.0,    0.0,   0.0,  2.0f32], // Translation along the Z-axis (2.0)
+                ];
+                
+                let perspective = {
+                    let (width, height) = target.get_dimensions();
+                    let aspect_ratio = height as f32 / width as f32;
+        
+                    let fov: f32 = 3.141592 / 3.0;
+                    let zfar = 1024.0;
+                    let znear = 0.1;
+        
+                    let f = 1.0 / (fov / 2.0).tan();
+        
+                    [
+                        [f * aspect_ratio, 0.0, 0.0, 0.0],
+                        [0.0, f, 0.0, 0.0],
+                        [0.0, 0.0, (zfar + znear) / (zfar - znear), 1.0],
+                        [0.0, 0.0, -(2.0 * zfar * znear) / (zfar - znear), 0.0],
+                    ]
+                };
+                let light = [-1.0, 0.4, 0.9f32];
+        
+                let params = glium::DrawParameters {
+                    depth: glium::Depth {
+                        test: glium::DepthTest::IfLess,
+                        write: true,
+                        .. Default::default()
+                    },
+                    backface_culling: glium::draw_parameters::BackfaceCullingMode::CullClockwise,
+                    .. Default::default()
+                };
+        
+                let view = camera::view_matrix(&[2.0, -1.0, 1.0], &[-2.0, 1.0, 1.0], &[0.0, 1.0, 0.0], zoom);
+        
+                target
+                    .draw(
+                        (&positions, &normals),
+                        &indices,
+                        &program,
+                        &uniform! { model: model, view: view, perspective: perspective, u_light: light },
+                        &params,
+                    )
+                    .unwrap();
+                target
+                    .draw(
+                        &vertex_buffer,
+                        &indices_2d,
+                        &program_color,
+                        &uniforms,
+                        &Default::default(),
+                    )
+                    .unwrap();
+        
+                target
+                    .draw(
+                        &vertex_buffer,
+                        &indices_2d,
+                        &program_texture,
+                        &texture_uniforms,
+                        &Default::default(),
+                    )
+                    .unwrap();
+                target.finish().unwrap();
+            }
             _ => {}
         }
-
-        let mut target = display.draw();
-        target.clear_color_and_depth((0.0, 0.0, 1.0, 1.0), 1.0);
-        let uniforms = uniform! {
-            matrix: [
-                [0.5, 0.0, 0.0, 0.0],
-                [0.0, 0.5, 0.0, 0.0],
-                [0.0, 0.0, 0.5, 0.0],
-                [0.5, 0.0, 0.0, 1.0f32],
-            ]
-        };
-
-        let texture_uniforms = uniform! {
-            matrix: [
-                [0.5, 0.0, 0.0, 0.0],
-                [0.0, 0.5, 0.0, 0.0],
-                [0.0, 0.0, 0.5, 0.0],
-                [-0.5, 0.0, 0.0, 1.0f32],
-            ],
-            tex: &example_image
-        };
-
-        let model = [
-            [0.0,    0.0,  -1.0,  0.0], // X-axis rotation (no change to X-axis values)
-            [0.0,    1.0,   0.0,  0.0], // Y-axis rotation (no change to Y-axis values)
-            [1.0,    0.0,   0.0,  0.0], // Z-axis rotation (no change to Z-axis values)
-            [0.0,    0.0,   0.0,  2.0f32], // Translation along the Z-axis (2.0)
-        ];
-        
-        let perspective = {
-            let (width, height) = target.get_dimensions();
-            let aspect_ratio = height as f32 / width as f32;
-
-            let fov: f32 = 3.141592 / 3.0;
-            let zfar = 1024.0;
-            let znear = 0.1;
-
-            let f = 1.0 / (fov / 2.0).tan();
-
-            [
-                [f * aspect_ratio, 0.0, 0.0, 0.0],
-                [0.0, f, 0.0, 0.0],
-                [0.0, 0.0, (zfar + znear) / (zfar - znear), 1.0],
-                [0.0, 0.0, -(2.0 * zfar * znear) / (zfar - znear), 0.0],
-            ]
-        };
-        let light = [-1.0, 0.4, 0.9f32];
-
-        let params = glium::DrawParameters {
-            depth: glium::Depth {
-                test: glium::DepthTest::IfLess,
-                write: true,
-                .. Default::default()
-            },
-            backface_culling: glium::draw_parameters::BackfaceCullingMode::CullClockwise,
-            .. Default::default()
-        };
-
-        let view = view_matrix(&[2.0, -1.0, 1.0], &[-2.0, 1.0, 1.0], &[0.0, 1.0, 0.0], zoom);
-
-        target
-            .draw(
-                (&positions, &normals),
-                &indices,
-                &program,
-                &uniform! { model: model, view: view, perspective: perspective, u_light: light },
-                &params,
-            )
-            .unwrap();
-        target
-            .draw(
-                &vertex_buffer,
-                &indices_2d,
-                &program_color,
-                &uniforms,
-                &Default::default(),
-            )
-            .unwrap();
-
-        target
-            .draw(
-                &vertex_buffer,
-                &indices_2d,
-                &program_texture,
-                &texture_uniforms,
-                &Default::default(),
-            )
-            .unwrap();
-        target.finish().unwrap();
     });
 }
 
-fn view_matrix(
-    position: &[f32; 3],
-    direction: &[f32; 3],
-    up: &[f32; 3],
-    zoom_factor: f32,
-) -> [[f32; 4]; 4] {
-    let f = {
-        let f = direction;
-        let len = f[0] * f[0] + f[1] * f[1] + f[2] * f[2];
-        let len = len.sqrt();
-        [f[0] / len, f[1] / len, f[2] / len]
-    };
-
-    // Calculate the normalized side vector (right)
-    let s = [
-        up[1] * f[2] - up[2] * f[1],
-        up[2] * f[0] - up[0] * f[2],
-        up[0] * f[1] - up[1] * f[0],
-    ];
-
-    // Calculate the normalized up vector
-    let s_norm = {
-        let len = s[0] * s[0] + s[1] * s[1] + s[2] * s[2];
-        let len = len.sqrt();
-        [s[0] / len, s[1] / len, s[2] / len]
-    };
-
-    // Calculate the final up vector
-    let u = [
-        f[1] * s_norm[2] - f[2] * s_norm[1],
-        f[2] * s_norm[0] - f[0] * s_norm[2],
-        f[0] * s_norm[1] - f[1] * s_norm[0],
-    ];
-
-    // Calculate the final position with zoom applied
-    let p = [
-        position[0] + f[0] * zoom_factor,
-        position[1] + f[1] * zoom_factor,
-        position[2] + f[2] * zoom_factor,
-    ];
-
-    // Construct the view matrix
-    [
-        [s_norm[0], u[0], -f[0], 0.0],
-        [s_norm[1], u[1], -f[1], 0.0],
-        [s_norm[2], u[2], -f[2], 0.0],
-        [
-            -s_norm[0] * p[0] - s_norm[1] * p[1] - s_norm[2] * p[2],
-            -u[0] * p[0] - u[1] * p[1] - u[2] * p[2],
-            f[0] * p[0] + f[1] * p[1] + f[2] * p[2],
-            1.0,
-        ],
-    ]
-}
